+++ BP/scripts/DoriosCore/machinery/index.js
@@ -13,3 +13,5 @@ export * from "./outputTracker.js"
 export * from "./itemContainers.js"
 export * from "./fluidContainers.js"
 export * from "./gasContainers.js"
+
+export * from "./itemEnergyStorage.js"